Install packages into base Conda environment
Instead of using a custom Conda environment to install packages, install into the base environment. Although the "proper" Conda way (as far as I understand) is to not install into the base environment, getting the various kernels to be aware of this environment is not straightforward. I imagine I will have to modify the kernelspec JSONs, or perhaps apply some other approach. In the meantime, I think installing into the base environment makes sense insofar as the "base" here is the breeder notebook itself. Users are still free to create and manage their own Conda environments as they see fit.